refactor application state logic: create useApplicationData hook
